Move the zoo into cmd
cmd is a better place for the zoo. It is a good oppotunity to test the
whole process of generation and centralizes two disparate but eerily
similar test areas.

This PR also adds a `README.md` on how to add a test case as well as how
to update a testcase.