Extract instructions from markdown
Parses the .md files for content and uses them to extract meaning that
can be carried forward into other files.

Creates a basic date parser

This date parser, unlike most others, doesn't just take a string like
"tomorrow" and return that date. It takes a date input as well which
allows it to be used to resolve dates from the past into the future or
in any other.

This means that we can now actually handle the example in the main
README.md file. Also updates the "simple" test in the zoo to be the same
as the README.md example.