Render parse errors from old files in latest
Parse errors are not super common, but should not be hidden.
For example you might type

tmrrow: do something

which has tmrrow instead of tomorrow which is not going to parse. Right
now this is handled by panicing in the parse (without any help to the
user on what triggered it) but we should emit an "errors" section at the
tail of the generated file.